+++ CHANGES.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * (wiki) Update methods with assert_exports
 * (wiki) add inner_class_name to setup page
 * (wiki) re-export cmdline help
+* (wiki) assert_called requires you pass defaulted param values in order to get a match with parameters.  Add an example.
 
 ## Features
 * `inner_class_name` option for editor, command line, and .gutconfig.json.